requirements.txt
streamlit
scikit-learn
pandas
nltk
newspaper3k
beautifulsoup4
requests
lxml
